cleaned up and removed deployment code that not used anymore
